Create utility function for image download
